backend/app/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.10
django-cors-headers==4.3.0
redis==5.0.1
django-redis==5.4.0
pandas==2.1.2
numpy==1.26.1
scikit-learn==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.1.0
django-filter==23.3
drf-yasg==1.21.7
celery==5.3.4
requests
soccerdata==1.8.7
